requirements.txt
Flask==3.1.0
supabase==2.11.0
python-dotenv==1.2.2
websockets==15.0.1
gotrue==2.12.4
realtime==2.28.3
postgrest==0.19.3
storage3==0.11.3
supafunc==0.9.4
